Response Messages

+++ src/main/java/com/mfs/merchantQR/repo/TblResponseMessageRepo.java
@@ -3,7 +3,7 @@ package com.mfs.merchantQR.repo;
 import com.mfs.merchantQR.model.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TblResponseMessageRepo extends JpaRepository<TblResponseMessage,Long> {
+public interface TblResponseMessageRepo extends JpaRepository<TblResponseMessage,Integer> {
 
     TblResponseMessage findByResponseMessageDescr(String responseMessageDesc);
